add: repository implement with mysql processing

+++ domain/model/training.go
@@ -3,6 +3,6 @@ package model
 // Training training struct
 type Training struct {
 	Date  string `json:"date"`
-	Count int64  `json:"count"`
+	Count int    `json:"count"`
 	Kind  string `json:"kind"`
 }
